[PATCH] ia64: correct ino_t to be 64 bits wide
This brings the kernel in sync with glibc (and all other 64-bit platforms
other than Alpha and S390x).  This fixes ustat() and breaks nfsctl_export,
but the latter is a deprecated interface anyhow (newer versions of
nfsutils use "exportfs" instead).

+++ include/asm-ia64/posix_types.h
@@ -10,7 +10,7 @@
  *	David Mosberger-Tang <davidm@hpl.hp.com>
  */
 
-typedef unsigned int	__kernel_ino_t;
+typedef unsigned long	__kernel_ino_t;
 typedef unsigned int	__kernel_mode_t;
 typedef unsigned int	__kernel_nlink_t;
 typedef long		__kernel_off_t;
